requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
openai>=1.0.0
chromadb>=0.4.18
tiktoken>=0.5.1
numpy>=1.24.3
pandas>=2.0.3
gradio>=4.0.2
spacy>=3.7.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2 